requirements.txt
numpy==1.15.0
pillow==5.2.0